examples/composio-demos/inbox-summarizer/requirements.txt
composio-core==0.7.12
google-genai==1.64.0
